+++ algoliasearch/src/main/java/com/algolia/search/saas/Query.java
@@ -349,10 +349,20 @@ public class Query extends AbstractQuery {
 
     private static final String KEY_FACET_FILTERS = "facetFilters";
 
+    /**
+     * Set the <b>deprecated</b> {@code facetFilters} parameter.
+     *
+     * @deprecated Use {@link Query#setFilters(String)} instead.
+     */
     public @NonNull Query setFacetFilters(JSONArray filters) {
         return set(KEY_FACET_FILTERS, filters);
     }
 
+    /**
+     * Get the value of <b>deprecated</b> {@code facetFilters} parameter.
+     *
+     * @deprecated Use {@link Query#getFilters()} instead.
+     */
     public JSONArray getFacetFilters() {
         try {
             String value = get(KEY_FACET_FILTERS);
@@ -368,10 +378,23 @@ public class Query extends AbstractQuery {
 
     private static final String KEY_FILTERS = "filters";
 
+    /**
+     * Filter the query with numeric, facet or/and tag filters.
+     * <p>
+     * The syntax is a SQL like syntax, you can use the OR and AND keywords. The syntax for the underlying numeric, facet and tag filters is the same than in the other filters:
+     * {@code available=1 AND (category:Book OR NOT category:Ebook) AND _tags:public date: 1441745506 TO 1441755506 AND inStock > 0 AND author:"John Doe"}
+     *
+     * @param filters a string following the given syntax.
+     * @return the {@link Query} for chaining.
+     */
     public @NonNull Query setFilters(String filters) {
         return set(KEY_FILTERS, filters);
     }
 
+    /**
+     * Get the numeric, facet or/and tag filters for this Query.
+     * @return a String with this query's filters.
+     */
     public String getFilters() {
         return get(KEY_FILTERS);
     }
@@ -1168,7 +1191,25 @@ public class Query extends AbstractQuery {
         return alternatives;
     }
 
+    private static final String KEY_RESPONSE_FIELDS = "responseFields";
 
+    /**
+     * Choose which fields the response will contain. Applies to search and browse queries.
+     * <p>
+     * By default, all fields are returned. If this parameter is specified, only the fields explicitly listed will be returned, unless * is used, in which case all fields are returned. Specifying an empty list or unknown field names is an error.
+     */
+    public
+    @NonNull
+    Query setResponseFields(String... attributes) {
+        return set(KEY_RESPONSE_FIELDS, buildJSONArray(attributes));
+    }
+
+    /**
+     * Get the fields the response will contain. If unspecified, all fields are returned.
+     */
+    public String[] getResponseFields() {
+        return parseArray(get(KEY_RESPONSE_FIELDS));
+    }
     // ----------------------------------------------------------------------
     // Parsing/serialization
     // ----------------------------------------------------------------------